created insertUser.go (unfinished)
