+++ backend/main.go
@@ -1,33 +1,33 @@
-package main
+// package main
 
-import (
-	"log"
-	"os"
+// import (
+// 	"log"
+// 	"os"
 
-	"github.com/VolunteerOne/volunteer-one-app/backend/database"
-	"github.com/VolunteerOne/volunteer-one-app/backend/models"
-	"github.com/VolunteerOne/volunteer-one-app/backend/server"
-	"github.com/joho/godotenv"
-)
+// 	"github.com/VolunteerOne/volunteer-one-app/backend/database"
+// 	"github.com/VolunteerOne/volunteer-one-app/backend/models"
+// 	"github.com/VolunteerOne/volunteer-one-app/backend/server"
+// 	"github.com/joho/godotenv"
+// )
 
-func main() {
-	database.Init() // Connect database
+// func main() {
+// 	database.Init() // Connect database
 
-	if os.Getenv("DB_MIGRATION") != "" {
-		models.Init()
-	}
+// 	if os.Getenv("DB_MIGRATION") != "" {
+// 		models.Init()
+// 	}
 
-	server.Init() // Start Server
-}
+// 	server.Init() // Start Server
+// }
 
-// Runs before main
-func init() {
-	// Load Environment Variables
-	err := godotenv.Load()
+// // Runs before main
+// func init() {
+// 	// Load Environment Variables
+// 	err := godotenv.Load()
 
-	if err != nil {
-		log.Fatal("Could not load environment variables")
-	}
+// 	if err != nil {
+// 		log.Fatal("Could not load environment variables")
+// 	}
 
-	log.Printf("Application starting on port: %v", os.Getenv("PORT"))
-}
+// 	log.Printf("Application starting on port: %v", os.Getenv("PORT"))
+// }
