+++ backend/models/initialization.go
@@ -12,6 +12,7 @@ type Model interface {
 // Register all models into this table
 var tables = []Model{
 	&Object{},
+	// &Users{},
 }
 
 func Init() {
